Core: Correct branch analysis truncation

+++ Core/MIPS/IR/IRCompBranch.cpp
@@ -42,7 +42,7 @@
 #define _SIZE ((op>>11) & 0x1F)
 #define _IMM16 (signed short)(op & 0xFFFF)
 #define _IMM26 (op & 0x03FFFFFF)
-#define TARGET16 ((int)((uint32_t)_IMM16 << 2))
+#define TARGET16 ((int)((uint32_t)(int)_IMM16 << 2))
 #define TARGET26 (_IMM26 << 2)
 
 #define LOOPOPTIMIZATION 0
